pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "behavioral-health-outreach-propensity"
version = "0.1.0"
description = "Production-grade propensity and uplift modeling pipeline for outreach prioritization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.1",
  "scikit-learn>=1.3",
  "xgboost>=2.0",
  "pyyaml>=6.0",
  "scipy>=1.11",
  "sqlalchemy>=2.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6"
]

[project.scripts]
outreach-train = "outreach_model.cli:run_train"
outreach-report = "outreach_model.cli:run_report"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
